renamed ra-full.py to ra.py
